fintrack/backend/requirements.txt
# ============================================================
# fintrack backend — Python dependencies
# File: /home/fintrack/fintrack/backend/requirements.txt
# ============================================================

# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication & security
python-jose[cryptography]==3.3.0   # JWT tokens
passlib[argon2]==1.7.4             # Argon2id password hashing
python-multipart==0.0.19           # Form data / file uploads
cryptography==43.0.3               # Encryption primitives

# Encryption — zero-knowledge model
PyNaCl==1.5.0                      # libsodium bindings

# Data processing — ported from your existing script
pandas==2.2.3
openpyxl==3.1.5

# Configuration
pydantic==2.10.2
pydantic-settings==2.6.1           # .env file loading
python-dotenv==1.0.1

# HTTP client — for future Plaid integration
httpx==0.28.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0                      # also used as async test client

# Utilities
python-dateutil==2.9.0
pytz==2024.2
